Implement Spark's variadic HASH() function
Summary:
Spark uses a variant of Murmur3; I've copied the hash implementation directly
from the Spark codebase, modified of course for C++.

The Spark algorithm for combining hashes is simple: initialize the hash to 42,
and then use the current hash as the seed for the next hash. NULLs are skipped.
This implementation processes one column at a time.

Differential Revision: D29743513

fbshipit-source-id: 2784e0959183559bf1bafec59e588a290986aa65