Rename ChecksumAlgorithm NULL option to NULL_
Summary:
gcc on linux (9.3 on Ubuntu 20.04) seems to have a problem with the
`NULL` option name and fails the compilation. Renaming the flag to NULL_
instead. Interestingly enough, all other usages of the flag I could find
already use "NULL_", so I'm a little confused about how this worked in the
first place:

> P445286158

Differential Revision: D30396818

fbshipit-source-id: 541fce87f3dabc837910377d1d293498f78cdce5