Simplified eval path - first cut
Summary:
First cut for a simplified expression evaluation path, which flattens
all vector before execution. The idea is to use this alternative path for tests
and to help isolate production issues. It will also help us understand how much
the current (complex) eval path buys us in terms of performance.
.
A few special form expression are not supported yet, in order to make this diff
a bit smaller.

Differential Revision: D30122151

fbshipit-source-id: a9847131a23e9f68ceb9bfd49f44186eeaa520f5