Add support for broadcast partitioning (#474)
Summary:
Pull Request resolved: https://github.com/facebookexternal/presto_cpp/pull/474

Broadcast repartitioning sends all the data produced by a query stage to all
the workers processing the next query stage. The implementation re-uses
PartitionedOutput operator configured to send all the data to a single
destination.

PartitionedOutputBuffer is modified to add is-broadcast flag. The total number
of broadcast destinations may not be known at the task start time. Hence,
there is a new method Task::updateBroadcastOutputBuffers which allows to
update the total number of broadcast destinations while the task is running.

PartitionedOutputBuffer holds on to produced data until the final number of
broadcast destinations is known so that it can send the data to destinations
that are added after the data was produced.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/81

Reviewed By: kgpai

Differential Revision: D30417033

Pulled By: mbasmanova

fbshipit-source-id: 91aaa03bf8ff44b25c580fcffaac4a962893dc1f