Add commented expr eval example binary
Summary:
Adding a binary containing a commented minimal example of the expr
evaluation engine. I would like to get some feedback on this binary, but the
main idea would be to provide a similar binary for other integration models and
extensibility model. Hopefully this can make it easier for Velox users to
understand the different usage modes, help highlight the external APIs, and
serve a reference for develpers integrating with Velox.

Reviewed By: kgpai

Differential Revision: D30452468

fbshipit-source-id: 7c0bb9169069eb83f8d54e7ae0ce947ee35ce5a9