Support different types for partial and final sum() aggregation. (#497)
Summary:
Pull Request resolved: https://github.com/facebookexternal/presto_cpp/pull/497

Pull Request resolved: https://github.com/facebookincubator/velox/pull/213

The existing problem is with Presto: for REAL (float) data it uses DOUBLE in partial aggregation and REAL in final.
Velox assumes that DOUBLE is used in both.
As a result we have discrepancy between plan and physical types and we throw error.
This diff allows sum() to have different types in partial and final aggregations.

Reviewed By: mbasmanova

Differential Revision: D30950310

fbshipit-source-id: 47ac21667e440c90e083234f1898dd48e3fbcedb