evalWithMemo() shouldn't bypass null optimizations
Summary:
For subexpressions where memoization is enabled, the eval follows the
evalWithMemo() path, and that ends up calling evalAll() directly, bypassing the
null optimizations in evalWithNull().
.
Results are mostly the same, except for the case where the expression
deterministically throws, and the null optimization would make the eval skip
the processing of this expression altogether.

Differential Revision: D31008389

fbshipit-source-id: f24e8896f8880bcd1d438360014fa9679c69c939