Add EnforceSingleRow operator
This operator enforces that input has no more than one row and returns that row
unmodified. If input is empty, this operator returns a single row with all
values set to null. If input has more than one row, this operator throws an
exception and aborts the query.

This operator is used in query plans that have non-correlated sub-queries, e.g.

	SELECT name
	FROM tpch.tiny.nation
	WHERE regionkey = (SELECT max(regionkey) FROM tpch.tiny.region)