Use std::copy instead of memcpy
Summary:
I'm debugging a crash in DecodedVector where indices_ is unexpectedly null.  In
mode/dev, the program is killed before I get to the relevant crash if the src
parameter is null even if the size is zero. std::copy doesn't do this.

Performance of std::copy should be identical to memcpy.

Reviewed By: pedroerp

Differential Revision: D30978509

fbshipit-source-id: 5e81d457bb69ac2b93882d37b31378ece187196f