Allow null rows in RowContainer::extractColumn (#259)
Summary:
RowContainer::extractColumn takes a list of pointers to rows from which to
extract values. This change adds support for having some of these pointers be
null in which case the corresponding values in the result vector are set to
null. This will be used in the implementation of the left hash join to extract
values for the build-side columns. For some of the rows there is no match on
the build side and we need to extract 'null' for these rows.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/259

Differential Revision: D31058600

Pulled By: mbasmanova

fbshipit-source-id: 6108c67abb37b538e3263286ed62aeb8413b6b53