Bug fix in consecutive reading in SelectiveColumnReader
Summary:
The bug is due to reader tries to read stuff beyond bufferEnd_. Boundary check is off here.

This error does not appear most of the time due to ASAN not enabled. However it will give WRONG results. When ASAN is not enabled it will only crash if the out of bound memory it tries to read is out of allocated mapped memory (chance is small). That is why this type of crash is not happening often.

Reviewed By: mbasmanova

Differential Revision: D31131027

fbshipit-source-id: 23bf36d75fb074650026c073aa8168017922c802