Add erase method to HashTable
Adds erase to HashTable for use in spilling selected groups from a
group by hash table.

A deleted entry is marked by a tombstone in the tags. This counts as
occupied for probing and as empty for inserting. If a group of tags
has at least one empty, then an erase is represented by setting the
tag to empty. If all tags are non-empty, the erase is represented by
setting the tag to a tombstone. This is done so that an erased entry
will not interrupt the probing when a tag is not found in a group of
tags. Note that probing stops when there is no hit and there is at
least one free in a group of tags.

- Adds eraseRows method to RowContainer.

- Fixes offset of probedFlag bit in RowContainer.