Implement cross join (#282)
Summary:
Introduce CrossJoinNode plan node to specify a cross join in a query plan. Cross
join takes two plan nodes for the left and right sides and an output type which
allows to reorder the input columns.

A query plan with a cross join executes using two pipelines: (1) build pipeline
processes the right side data, collects all data in a list of vectors and makes
it available to the probe pipeline; (2) probe pipeline processes the data on
the left side in a streaming fashion and combines it with the right side data.

CrossJoinNode is translated into two operators: CrossJoinProbe and
CrossJoinBuild. CrossJoinProbe operator becomes part of the probe pipeline.
CrossJoinBuild operator is installed as the last operator of the build side
pipeline. The output of the CrossJoinBuild operator is a list of build side
vectors which CrossJoinProbe operator gets access to via CrossJoinBridge.

CrossJoinProbe wraps probe and build side vectors in dictionaries to represent
repeated values without copying.

TPC-H q11 and q17 are now succeeding.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/282

Reviewed By: oerling

Differential Revision: D31141606

Pulled By: mbasmanova

fbshipit-source-id: 1974995a4015b3949cf594270c150cb601ee951b