Fix key channel indices in PartitionedOutput
Key channel indices are relative to the channels in the input
RowVector. The composition and order of the output may be different
and may not include the partitioning channels. Therefore interpret
'keyChannels_' in terms of the original input and do the rest of the
processing in terms of the rearranged input vectors stored in
'output_'.

Rename input to output where applicable. Add a test that drops the
partitioning column and triggers and error before the fix.