Track size of row in RowContainer
Adds a uint32_t to the row for tracking width of out of line variable
length data. This is only added if there are variable length fields or
accumulators.

This allows controlling the output batch size for hash join or group
by and for getting predictable size chunks of data for spilling.

Follow-up PRs will add setting this in hash build, group by key insert
and when growing variable width accumulators.

Sets the optional probed flag to 0 in initialNulls_' of a new row.