Use AsyncDataCache and read-ahead with Hive
Uses AsyncDataCache and ScanTracker with Hive if the MappedMemory
given to the reader is a AsyncDataCache. If this is a regular
MappedMemory, uses BufferedInput with the given DataCache, if any.

Sets most tests to use AsyncDataCache as MappedMemory also if no
affinity.

Tests TableScan with both AsyncDataCache and regular MappedMemory +
SimpleLruDataCache.