Cleanup FieldAccessTypedExpr
FieldAccessTypedExpr Represents one of two things:
	- a leaf in an expression tree specifying input column by name;
 	- a dereference expression which selects a subfield in a struct by name.

In the first case, it doesn't have any inputs. In the second case, its input is
an expression that returns a struct.

This change is to replace a single constructor that takes a list of inputs and
assert there is exactly one element with two constructors: one taking no input
expression and another taking exactly one input expression.