Fix integer overflow in resizing
Summary:
# Problem
When we resize a `FlatVector`, we resize the internal buffer with the byte size required to store the given # of elements. To be exact:
First, Get the byte size required.
```
vector_size_t minBytes = BaseVector::byteSize<T>(size);
```
Second, Resize the buffer
```
values_->setSize(minBytes);
```

The problem was:
- `setSize()` accepts an **`uint64_t`**.
- `byteSize()` returns `vector_size_t`, which is **`int32_t`**.
- With a big enough `size`, `byteSize()` hit integer overflow.
- __And the overflowed `int32_t` gets implicit-casted to `uint64_t`__.

Resulting we get an exception like this:
```
Error Source: RUNTIME
Error Code: INVALID_STATE
Reason: (18446744071562069392 vs. 3221225376)
Retriable: False
Expression: size <= capacity_
Function: setSize
```

So the limit that a `FlatVector` could hold was way-lower than the `int32_t` limit, and the exception was confusing.

# Solution
`byteSize()` must return `uint64_t` which is what we use for buffer size. `vector_size_t` is not the right type for the case. I changed it and now it looks working.

I did not add safe-multiplication, in `byteSize()` function. We would hit memory problems way before that to be a problem.

Differential Revision: D31528591

fbshipit-source-id: d1e31b4eff51e5866eb2dd3c21d9fa1bcc347b7e