Implement FixedSizeArray type. (#366)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/366

This is the first step to adding tensor support to Velox.  Here we add support for fixed size arrays.

Purpose: support implementation of Tensor types in TorchArrow (https://docs.google.com/document/d/1QIrcJinkO8yCAVDQGwElPh2VUNXrHrfuhOkkq2m5Ohg/edit)

In this diff we add the type itself.  A FixedSizeArray is an array that is
constrained to always have member elements of a given fixed size.  Its
equivalent in Arrow type system is Fixed-Size List:

https://arrow.apache.org/docs/format/Columnar.html#fixed-size-list-layout

We implement FixedSizeArrayType as a subtype of ArrayType so that anywhere an
ArrayType can be used, a FixedSizeArrayType can be provided.  Currently neither
Presto / Spark have a notion of fixed size array, so when passing this type
over the wire it maybe become a general variable width array.

Reviewed By: pedroerp

Differential Revision: D31164681

fbshipit-source-id: b075112f423f9d401158318aa7d69e7ba0c2dbb7