Add GroupTracker for file group access tracking
GroupStats holds process-wide access stats for column access in each
file group. A file group is typically a Hive partition of a
table. Different partitions will have different access frequencies for
different columns. For each file group we estimate the number of
distinct files and the number of splits per file. For columns, we
track the references and actual reads.

This information gives us the sizes of the hot columns and allows
deciding which ones should be staged on SSD. We order group/column
pairs on a score and qualify the top k GB worth of columns to be
written to SSD at the next opportunity.

Only columns that are actually referenced by queries are
recorded. Unlike ScanTracker, this does not differentiate between
different streams of one column.  If one stream will be cached on SSD,
then all the related streams should also be cached.

The number of columns can be very large. Therefore the tracker has a
size limit and removes tracking data for columns if there are no
accesses. Column access stats can be periodically decayed so as to
deemphasize old activity. The frequency and amount of decay is up to
the caller.