Fix like function when no escape char specified
The 'like' function logic had a bug when escape char
was not specified in its invocation. In this situation,
the escape char was left uninitialized and could pick up
any value. If that random value matched a char in the
'pattern' parameter then it resulted in failures with error
'Escape character must be followed by ...'

This change fixes the bug by tracking using a boolean if indeed the
escape char was specified in the invocation, and modifying the logic
accordingly.