Make ScalarFunction check call()'s full signature (#479)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/479

Previously, ScalarFunction was dispatching execution to simple functions
by simply checking which method was available (call/callNullable/callAscii).
With this change, we now check the full method signature (including return type
and parameters), which enables us to template the call method directly, in
addition to mix and match call, callNullable and callAscii based on the input
parameters.
.
It will also allow us to simplify the way simple functions are specified, in
the next diff/PR.

Reviewed By: zzhao0

Differential Revision: D31831081

fbshipit-source-id: 670070062f042ba4c761023fa576c56e7a640c27