Modify enum values in TypeKind
The enum values in TypeKind are contiguous between the simple scalar and
Complex types. This was resulting in odd looking code when adding new
scalar types.

This PR leaves some enum values space between the simple and complex
types so that new types can be easily added without moving around the
enum values.