Re-fix protobuf generation
My original fix (D30134167) seems to have been partially reverted; this
PR just re-applies the core changes along with some minor style fixes to
the CMake files.

The build was failing sometimes because velox_dwrf_int_decoder_benchmark
was missing a dependency on velox_dwio_dwrf_common, which transitively
depended on the proto headers. This would sometimes result in this
target being built before the proto headers were generated.

`include_directories(${CMAKE_BINARY_DIR})` at the top of the CMake file
in this directory was masking this. Now we should once again get
deterministic failures if a dependency is missing.