Explicit null result for String to Int cast (#476)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/476

Current for converting String to Int, Velox uses folly::to conversion method that uses exception for error handling. The engine catches those exception and either fail or converts to null. In order to reduce the exception cost, as it was evident in some queries, this diff replace folly::to with our own stringToInteger conversion method and return error as an explicit input called nullOutput.

Reviewed By: funrollloops

Differential Revision: D31769513

fbshipit-source-id: 06160ba45b5d4e55ab11ea073691abd763985a6a