Add benchmark for Presto IN expression (#516)
Summary:
The benchmark shows that IN expression with large number of values is very
slow. This is because IN is implemented as a function with variable number of
arguments. IN with 1K values is represented as an expression tree with one
node for the IN, one for the column and 1K nodes for values. Each value is
wrapped in a ConstantExpr which gets evaluated for each batch of rows making new
shared pointers to constant vectors.

A follow-up PR will optimize IN expression to take values as a single array.

```
============================================================================
velox/functions/prestosql/benchmarks/InBenchmark.cpprelative  time/iter  iters/s
============================================================================
fastIn                                                       4.38ms   228.06
in                                                34.65%    12.65ms    79.02
fastIn1K                                                     4.02ms   248.83
in1K                                               1.84%   218.08ms     4.59
============================================================================
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/516

Differential Revision: D31997776

Pulled By: mbasmanova

fbshipit-source-id: 40da127b8b9a46181c26e2cf2eb0f975f05fc6f4