Fix IN predicate for Spark
Summary:
Vector function that has a single argument is guaranteed to receive its input as
flat. Vector function that has multiple arguments all but one of which are
constant is guaranteed to receive its not-necessarily-constant argument as flat
or constant. However, in this context "constant" doesn't mean vector of
constant encoding. Instead, it means that the vector has same value for all
rows in scope, e.g. BaseVector::isConstant(rows) is true.

Differential Revision: D32197009

fbshipit-source-id: 71ed87cfa3446efe73e9960edc817b32256e7b46