Add delay for checking postconditions of DriverTest
Some postconditions of completing or cancelling a query are realized
slightly after the last thread acknowledges the stop or a cursor is
marked as being at end. Specially the latter case is not practical to
be fused in the same critical section that marks the result set as
complete. We add a macro to test the postcondition with a delay so as
to reduce test flakiness.