Test Row Group Skipping in SelectiveColumnReader
Adds a test with column values that allow row group (stride)
skipping.

Fixes seekToRowGroup for structs.

Limits use of seekToRowGroup() to seeking on top level columns
alone. The row numbers of the column must be aligned with the top
level row numbers for this to work.

Adds test coverage for structs inside lists and maps.

Fixes addValue for string dictionary ValueHook. This must not
reference the in dictionary bitmap because this is aligned to
non-nulls for the bulk path and to values including nulls for the
value at a time path. Instead, use the dictionary index to choose
between stripe and row group dictionaries.