Quantize MemoryUsageTracker Update
Modifies MemoryUsageTracker to automatically make a quantized
reservation to cover allocations so that updates are rounded up and
propagated upwards in the tracker tree in bulk. Propagating at every
update causes the tracker tree to take up to 10% of profile due to
cache coherence traffic from all threads hitting one top tracker.

Rounds allocations to the next MB if under 16MB totake and 4 or 8 MB
if larger. Non-leaf trackers summarize the reservation at this
granularity.