Add MergeJoin operator (#638)
Summary:
MergeJoin operator assumes that data on both sides of the joins is sorted on the
join keys and streams both join sides looking for matching rows and emitting
results.

This PR adds support only for the inner merge join with no additional filter.
Other joins types as well as filter support are TODOs for future PRs.

MergeJoinNode plan node represents a merge join in a query plan. It has the same
properties as HashJoinNode, e.g. join type, join keys, optional filter and
output columns.

To execute a plan with a merge join, Velox creates two separate pipelines. One
pipeline processes the right side data and puts it into JoinMergeSource. The
other pipeline processes the data on the left side, joins it with the right
side data and continues execution as specified by downstream plan nodes.
MergeJoinNode is translated into MergeJoin operator and a CallbackSink backed
by JoinMergeSource. MergeJoin operator becomes part of the left-side pipeline.
CallbackSink is installed at the end of the right-side pipeline.

JoinMergeSource is used to coordinate data transfer between right side
(producer) and left side (consumer) of the join.

<img width="800" alt="merge-join-pipelines" src="https://user-images.githubusercontent.com/27965151/142435947-c4bfb62e-a079-4ab9-bca6-311a775f5448.png">

The main complexity in this implementation comes from the need to support cases
when matching rows span 2 or more batches of input on the left, right or both
sides.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/638

Reviewed By: oerling

Differential Revision: D32530846

Pulled By: mbasmanova

fbshipit-source-id: e1649ebde47125297fdd5c6d5d7135e733e7b386