Port array min/max to simple function interface (#635)
Summary:
For primitives benchmarking shows that we are close enough perf wise (16% slower)
with some planned tasks to match the performance.

One thing to note is that the new udf authoring interface
(the non macro based) was not friendly for this case so I used the
old one. The new interface requires explicit creation of 18 partially instantiated aliases before
registration.
I tried to use member struct alias to avoid that but c++ does not template deduction when the
argument is member alias. i.e.
registerFunction< X<..>::type, ..>(); // not allowed.

The string varient can use the 0-copy optimizations once the current work on it is done.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/635

Reviewed By: mbasmanova

Differential Revision: D32518343

Pulled By: laithsakka

fbshipit-source-id: e40ef85a654f15bf103817201629ebbcee7240a3