Adjust TableScan batch size by DataSource row size estimate
A constant batch size of 1K rows has been seen to lead to timers and
other per-batch overhead going to up to 10% of profile for narrow and
sparse workloads. Furthermore some of this expense involves mutexes,
for example inside retrieving CPU thread time, surprising as this may
seem. This also increases safety for very wide tables, e.g. maps of
10K keys.

We aim at a target batch size of up to 10MB, which for 100 threads
would be 1GB * ~5 operators per pipeline = ~5GB, which is a reasonable
figure for intermediate result size.

A more portable implementation could be to compare raw input size to
processed input rows but this would undercount sparsely accessed
columns, be relative to compressed size and would count in extra reads
from coalescing, which would make this quite unusable.

The code is covered by TableScanTest.* as is.