FIX: Disable vector reuse. (#667)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/667

Many functions/components are implemented with the assumption
that not-null do not need to be set, since nulls are initialized
with all not null.

vector re-use invalidates that assumptions since it does not
reset the nulls.
This diff disables the vector re-use and makes ExprSet::eval
clears results vector if used before.

prepareFlatResultsVector was defined twice, one
definition is kept.

Differential Revision: D32677187

fbshipit-source-id: d9c03ac4bcec9ade70a629e10062a864099ceed1