Do not use MapVector::canonicalize outside of construction
Comparing or serializing data to RowContainer must not mutate the
source data. Maps must be compared and serialized in key order but
canonicalize cannot be used for this because this will read to flaky
crashes if different threads handle one MapVector, which can happen
for example with local exchange.

Sorts MapVector element indices by the key as needed when serializing
to RowContainer or comparing for non-equality.

Uses a commutative hash mix for maps since the key order is uncertain
and we do not wish to create key order for hashing.

Checks that canonicalize is only used on singly referenced
MapVectors. This operation is still useful at construction time for
removing duplicates.

Coverd by exec/tests/RowContainerTest and vector/tests/VectorTest.cpp.