Throw on ambiguous function definitions (#689)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/689

Today the Registry class will overwrite any entry in the CreatorMap if a duplicate key is used.

For UDF's in ScalarFunctionRegistry and AdaptedVectorFunctionRegistry this will become more error prone with the introduction of support for VariadicArgs in the function signature.  In order to facilitate UDF lookups, I plan to modify FunctionKey to treat compatible signatures  as equal (e.g. VariadicArgs<Varchar> will match Varchar, Varchar).

This will make it much easier to accidentally overwrite another function in the Registries.  To help catch this early, I've added a check in the Register function that throws an exception if an entry already exists in the Registry with the same key.

This will also impact the other Registries (DeserializationRegistryType and AggregateFunctionRegistry) but I think this makes sense (overwriting an existing key is probably a mistake).

Reviewed By: mbasmanova

Differential Revision: D32780498

fbshipit-source-id: bba1d9656bd90d4376eb370db08866453ca5b4bd