Better view types iterator implementation. (#649)
Summary:
The previous iterator design maintains element of the traversed type that it updates its index as it moves,
and returns a reference/pointer to that element when it's accessed.

The problem with that is that it++ invalidates reference to inner elements obtained through it.
The following code demonstrates the problem:
```
    auto it = arrayView.begin();
    auto & v = *it;
    cout<<v;
    it++;
    cout<<v;
    return 0;
```
This diff solve that problem and simplify the iterator. Utilizing the fact that the if the returned type
of operator-> is not a pointer the compiler will apply -> recursively until a pointer is returned.

All complex types hence are overloaded with -> returning this,  for OptionalValueAccessor since ->
has its own meaning(unsafe pointer value access). a type OptionalValueAccessorPtr was created .

I addition to that this diff does the following:

- the lazyValueAccessor for the map key, is not used anymore. instead the value is eagerly materialized at construction
(if its complex, its materialization of view type). Since the construction happen inside the loop iteration now, and since it's
never written again, if it's never used the compiler should be able to remove the redundant writes.

- finally this diff also add -> operator for the std::optional like object with same meaning as std::optional.

### performance :
```
============================================================================
velox/functions/prestosql/benchmarks/MapInputBenchmark.cpprelative  time/iter  iters/s
============================================================================
mapSumVectorFunction                                         8.18ms   122.25
mapSumSimpleFunction                             112.84%     7.25ms   137.95
nestedMapSumVectorFunction                                  24.85ms    40.25
nestedMapSumSimpleFunction                       111.45%    22.29ms    44.86
nestedMapSumVectorFunctionMapView                129.27%    19.22ms    52.03
============================================================================

============================================================================
velox/functions/prestosql/benchmarks/ArrayMinMaxBenchmark.cpprelative  time/iter  iters/s
============================================================================
fastMinInteger                                             543.87us    1.84K
vectorMinInteger                                  80.40%   676.42us    1.48K
vectorMinIntegerNoFastPath                        55.91%   972.75us    1.03K
simpleMinIntegerIterator                          55.14%   986.27us    1.01K
simpleMinInteger                                  52.75%     1.03ms   969.96
============================================================================
```

Pull Request resolved: https://github.com/facebookincubator/velox/pull/649

Reviewed By: funrollloops

Differential Revision: D32596654

Pulled By: laithsakka

fbshipit-source-id: 0146186d9c662f7b0f40184e3d282d626559bdfc